feat: refactor for packaging
