+++ arkhi/server/server.ts
@@ -1,19 +1,14 @@
-// Note that this file isn't processed by Vite, see https://github.com/vikejs/vike/issues/562
-
-import express from "express";
 import compression from "compression";
+import express from "express";
 import { renderPage } from "vike/server";
-import { root } from "./root.js";
+import { applyMiddleware } from "./middleware.js";
 import * as trpcExpress from "@trpc/server/adapters/express";
-import { appRouters } from "../api/index.js";
-import { createContext } from "./context.js";
 
 const isProduction = process.env.NODE_ENV === "production";
 
-startServer();
-
-async function startServer() {
+export async function startServer() {
 	const app = express();
+	const root = process.cwd();
 
 	app.use(compression());
 
@@ -31,13 +26,15 @@ async function startServer() {
 		app.use(viteDevMiddleware);
 	}
 
-	app.use(
-		"/trpc",
-		trpcExpress.createExpressMiddleware({
-			router: appRouters,
-			createContext,
-		})
-	);
+	applyMiddleware(app);
+
+	// app.use(
+	// 	"/trpc",
+	// 	trpcExpress.createExpressMiddleware({
+	// 		router: appRouters,
+	// 		createContext,
+	// 	})
+	// );
 
 	app.get("*", async (req, res, next) => {
 		const pageContextInit = {
