+++ package.json
@@ -5,7 +5,7 @@
     "dev": "npm run server",
     "prod": "npm run build && npm run server:prod",
     "build": "vite build",
-    "server": "ts-node ./arkhi/index.ts",
+    "server": "ts-node server.ts",
     "server:prod": "cross-env NODE_ENV=production ts-node ./arkhi/index.ts"
   },
   "dependencies": {
@@ -13,7 +13,6 @@
     "@trpc/client": "^10.37.1",
     "@trpc/server": "^10.37.1",
     "@types/compression": "^1.7.2",
-    "@types/express": "^4.17.14",
     "@types/node": "^18.11.9",
     "@types/react": "^18.0.8",
     "@types/react-dom": "^18.0.3",
@@ -28,10 +27,13 @@
     "superjson": "^1.13.1",
     "ts-node": "^10.9.1",
     "typescript": "^4.9.4",
-    "vite": "^4.0.3",
     "vike": "^0.4.92",
+    "vite": "^4.0.3",
     "zod": "^3.22.2"
   },
   "license": "MIT",
-  "type": "module"
+  "type": "module",
+  "devDependencies": {
+    "@types/express": "^4.17.18"
+  }
 }
